Terminadas validaciones y funcionalidad

+++ db/migrate/20170310013457_add_price_product_to_products.rb
@@ -1,5 +1,5 @@
 class AddPriceProductToProducts < ActiveRecord::Migration[5.0]
   def change
-    add_column :products, :price_product, :integer
+    add_column :products, :price_product, :bigint
   end
 end
